Fix seven defects found in self-review, and close test gaps
Correctness
-----------
* Five subsystem getters returned a const reference from a UFUNCTION.
  That is accepted by some engine versions and rejected by others, which
  is a total build failure on the wrong one. They now return by value;
  the 80-200 byte copies are taken a few times per frame against a tick
  whose expensive branch is a cubemap render. Callers that bind the
  result to a const& are still correct via lifetime extension, and now
  hold a stable snapshot rather than a live reference.

* FArchSkyReplicatedState quantised the time flow rate into an int16 at
  a scale of 100, saturating at 327.67 h/s - well below the +/-600 the
  clock permits. Every playback speed above that would have put clients'
  suns somewhere the server's was not. Widened to int32; dropping the
  scale instead would have quantised the real-time rate to zero.

* The ViewModel's throttle had its guard inverted: an interval of 0,
  meaning "no throttle", would have stopped the text updating entirely
  rather than updating it every frame.

* Weather tile thumbnails were never populated, so the tiles the brief
  asks for had no image to draw. Added
  UArchSkySubsystem::GetWeatherPresetThumbnail and wired it through.

* The sun-path visualiser registered against USceneComponent. Visualisers
  are keyed by class name, one per key, so it would silently replace -
  and be replaced by - any other plugin doing the same. Added a one-line
  UArchSkyDirectorRootComponent to give the registration a key of its own.

* Loop mode did not hold the date. A window ending at midnight crossed
  it and the date rolled, so "loop this day" quietly became "loop
  forward through the calendar". The date is now restored on a wrap,
  using the previous tick's recorded date so a year boundary needs no
  arithmetic.

* Dragging a slider fought the pointer: the push throttle meant the next
  refresh wrote a value a frame old straight back into the handle the
  user was holding. Write-back now skips the dragged handle, with mouse
  capture hooked on all four sliders.

Performance
-----------
* The ViewModel rebuilt every FText on every state change - which, while
  the clock runs, is every frame: roughly twenty FText::Format calls
  each doing a localised number format, easily dearer than the whole
  astronomy layer. Split into a cheap numeric half that still runs every
  frame and a formatted half throttled to 10 Hz, never throttled while
  paused so a scrub always lands. The widget skips its SetText block
  unless FormattedRevision moved, since SetText invalidates Slate layout
  whether or not the text differs.

* Resolving the city name rebuilt a sixteen-entry array of FText-bearing
  structs every frame. Cached against the coordinates it was resolved
  for.

Tests
-----
Added ArchSky.State: weather blending (exact endpoints, alpha clamping,
linear-space colour lerp, discrete fields selecting rather than
interpolating across a hundred alphas, short-way bearing blending, and
the symmetry of the change metric the recapture throttle depends on),
state sanitisation (clamping, wrapping, idempotence, and a 200-case
hostile-input sweep) and replication quantisation - including the
assertion that catches the int16 saturation above.

24 automation tests total.

+++ Plugins/ArchSkySystem/Source/ArchSkyRuntime/Public/Core/ArchSkyState.h
@@ -113,9 +113,18 @@ struct ARCHSKYRUNTIME_API FArchSkyReplicatedState
 	UPROPERTY(BlueprintReadOnly, Category = "ArchSky|Networking")
 	uint16 Year = 2026;
 
-	/** Simulated hours per real second, quantised to 1/100. Lets clients predict forward. */
+	/**
+	 * Simulated hours per real second, quantised to 1/100. Lets clients predict forward.
+	 *
+	 * ARCH NOTE: int32, not int16. The clock permits +/-600 h/s, which at a scale of 100 is
+	 * +/-60000 - nearly double what an int16 can hold, so the fastest playback speeds would
+	 * silently saturate at 327.67 h/s and every client would predict the sun in the wrong
+	 * place. Dropping the scale to 1/10 would fit, but then the default real-time rate of
+	 * 0.0167 h/s would quantise to zero and time would appear frozen on clients. Two extra
+	 * bytes at 2 Hz is the obviously correct trade.
+	 */
 	UPROPERTY(BlueprintReadOnly, Category = "ArchSky|Networking")
-	int16 QuantisedTimeFlowRate = 0;
+	int32 QuantisedTimeFlowRate = 0;
 
 	/** Weather blend alpha quantised to 1/255. */
 	UPROPERTY(BlueprintReadOnly, Category = "ArchSky|Networking")
